Se agregan roles

+++ convex/files.ts
@@ -31,7 +31,7 @@ async function hasAccessToOrg(
   if(!user){return null;}
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some(item=>item.orgId===orgId) || user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) {
     return null;
@@ -113,6 +113,12 @@ export const deleteFile = mutation({
     if (!access) {
       throw new ConvexError("No access to file");
     }
+    const isAdmin = access.user.orgIds.find(org=>org.orgId===access.file.orgId) ?.role==="admin";
+
+    if(!isAdmin){
+      throw new ConvexError("You have no admin access to delete")
+    }
+
     await ctx.db.delete(args.fileId);
   },
 });
